If the feedback was not approved after its report, it is disabled instead of being physically deleted from the database
