+++ controller/report_feed.go
@@ -52,7 +52,6 @@ func ReportFeed(c *gin.Context) {
 		return
 	}
 	admins, err := db.GetAllAdmins()
-	_, err = db.GetAllAdmins()
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"Coudn´t get the admins": err.Error()})
 		return
@@ -70,7 +69,7 @@ func ReportFeed(c *gin.Context) {
 			c.String(http.StatusBadRequest, "An error has ocurred sending the email "+err.Error())
 			return
 		}
-	}
 
-	c.String(http.StatusCreated, "Feedback reported.")
+		c.String(http.StatusCreated, "Feedback reported.")
+	}
 }
